+++ src/client/index.ts
@@ -1,6 +1,7 @@
 import { HMR_HEADER, HMRPayload } from "../hmr";
 import { ErrorOverlay, overlayId } from "./overlay";
 import "./click-handler";
+import "./css-devtools";
 import * as RefreshRuntime from "./refresh-runtime";
 
 RefreshRuntime.injectIntoGlobalHook(window);